+++ sporktris.cpp
@@ -122,7 +122,7 @@ uint8_t Tetromino::Rect::center_x() {
   return (x + x2) / 2;
 }
 
-Sporktris::Sporktris(Display& disp, Controller *controllers, int controller_count):
+Sporktris::Sporktris(Display& disp, Controller *controllers, uint8_t controller_count):
   InputProcessor(controllers, controller_count), disp(disp) {
   button_conf[Controller::Button::b] = { .initial = 500, .subsequent = 500};
   button_conf[Controller::Button::a] = { .initial = 500, .subsequent = 500};
